Cleaned up building strings, removed external dependency.
I attempted to clean up the construction of the status messages by iterating over the items inside the dict we get cleaner code, fewer lookups and because `.items()` returns a view over the contents of the dict, we eliminate a race condition where a thread could update its state in the `threadStateDictionary` between the if statement and the string construction for the `print`.

The package `netaddr` is an external dependency which makes this slower to get up and running.  By using the python stdlib package `ipaddr` we eliminate that and remove the need to do the dance to remove the broadcast address by using the `.hosts()` method of the `ip_network` class.

I would recommend running this version before merging to ensure I didn't introduce any bugs.  I don't have the environment you have setup and its pretty clearly set up to match a specific environment.